It seems that the correct way to handle new variables is a judicious combination of EG(uninitialized_zval_ptr) and new allocations using ALLOC_INIT_ZVAL. In particular, we have to be very careful that we dont overwrite EG(uninitialized_zval_ptr) with data, and that we be aware that sometimes its refcount is greater than 1, and so check for it explicitly.
When cloning vars being passed to function calls, it seems it is necessary to decrement their refcount, in order to get the same results as the interpreter. I suspect this is due to the interpreter using a different call mechanism to us, and that is separates the zvals at a later point than us (this is conjecture, btw).

If a new variable is created when calling a function, make a note at runtime, and decrement the refcount after the return. Otherwise we will leak memory.