Using write_array_index for separating and referrence_array_index is incorrect, since we're supposed to write over the lhs, whereas write_array_index takes the lhs into account (as it should for non-referenced assignments, but not reference ones).
I inlined the functions, and removed the interactions with the lhs.

At this point, a fair bit of refactoring on libphc should be done. There's a lot of duplicate code, and this is going to start to hurt soon, with the volume of code that's already there.

In the code generator, fixed a warning.