My workflow uses indent to nicen the generated code before I compile it. As a result, I keep libphc.cpp indented using the defaults of the indent program. This commit re-indents liphc.cpp.
A number of places checked if a variable exists, when its being taken out of a symbol table. Its also necessary to check that it != EG(uninitialized_zval_ptr), otherwise we overwrite other variables, which results in very subtle errors.

Add casting to libphc.cpp and make Cast_array use it.


Add get_constant, which wraps around zend_get_constant. 

In Generate_C, process_rhs() will be run between fetching the data and writing it back. Use this for cast_array. As a result, we no longer use generate_epilogue.