completed the first part of removing the rhs memory allocations, and instead fetching the lhs before working on the rhs.
So far it only works for Bin_ops, lowering the running time to 8.5-9 seconds. 

This disables the Str_cat optimization, since it is no longer necessary.

It also adds a small amount of scaffolding (see fixed ()), which is to removed when finished.


Removed process_rhs, since it's no longer required to do the cast properly.