Remove some unnecesarry steps from ht_* functions.
We dont need to allocate a new zval for double conversions, a C level cast is fine.

We dont need to make a copy for string indexing, since its already a string.

We dont need a zend_symtable_* call for NULL, just a zend_hash_* call is fine.

Some indentation and spellings and TODO also included.