Overall, the aim was to have an interpret_hir test, with the actual HIR, as opposed to the AST version. The main change is to modify the Pass_manager to support a HIR as well as an AST. Since the AST and the HIR dont have a common base class, this is a bit difficult. I ended up wrapping them both in an IR class, which is dirty but manageable. I put in an issue about this, and when it rectified, it can be redone nicer.
Since the HIR_unparser doesnt support the --udump, since there is no loop construct in the HIR, it makes it a little bit difficult to test. I might involve an HIR_to_AST fold, but it isnt done yet. However, the necessary changes are made to the pass manager to support it.