Instead of a single pass queue, have pass queues for each IR. There are only a limited number of IRs, so this doesnt lead to too much duplication.
We now have a list of queues, which right now contain only AST and HIR queues. We iterate through the queues (folding when necessary), running the passes on each. phc.cpp was updated for the new API.

This also does some cleaning up. Folding is moved to IR, and so is hidden from phc.cpp. Plugin_pass gets its own .cpp file, and List_passes.h is gone. --list-passes now prints out which IR it uses, and the test is updated for it.