Add the MIR. This does enough to pass the quick tests. The MIR is the same as the old HIR, and the HIR is now a copy of the AST. AST_to_HIR is now a simple transform that creates a copy of the same structure, and ast_to_hir/ has mostly been moved to hir_to_mir/.
The plugins have been updated, and at the same time made more generic (mostly using templates) so that there is now no code duplication. Only the xml_roundtrip doesnt work, since we cant currently parse MIR and HIR. Thats a task for another day.

In the future, lowerings will be done in two parts, for the MIR and for the HIR. The HIR needs to be simplified, as does the MIR. We need to be able to fold the HIR back to the AST, and reuse the AST_unparser.