Fix a number of bugs related to unsupported features, and other minor bugs.
cb_mir doesnt work with iteratoing across an array that uses "" as a key. This is due to bugs with ArrayIterator. Since its an SPL problem, and only comes up on cb_mir test (which we really dont care about, its just an internal test, we ignore).

There is an error when you try to take a reference to a literal. Since we shred, we lose that error. Mark this unsupported.

Uppered dump doesn'y work for the HIR. This isnt a huge problem, so remove the test, add an error, and add a test case for the error.