Add more comments.
Fix the transfer function for live_in.

Add error handling, so that we fail if we havent handled any part of the CFG.

We dont want to propagate the vars themselves, as they have different IDs for each instance. We want to propagate the var_names. Infer them with used_var and defined_var.