Calypso cannot encode a non-deterministic algorithm. This means that I can't mark a cfg_node as dead, and then run the analysis again to find more dead nodes. Obviously, this would kill the whole point of optimizing, if you cant iterate.
So there is a nice solution. Each CFG is versioned. If anything is marked dead, then a new version is generated without that node. Optimizations are re-run on this new version. It works by adding the new version as a separate DB. Calypso then picks up on the new version. If there are no dead nodes, then we are finished.

So, as long as each of these optimizations is monotonic, this versioning will handle the fixpoint-ing, and my analyses will terminate.

There are some bugs to be figured out here, notably branches and gotos need to be fixed. Fortunately, these will only need to be fixed once, and the details can be abstracted away from every other part of the optmizations.