Reorganizes do_optimize into smaller files with a single purpose each. do_versioning handles iteration and fixpointing, and linearize_cfg handles cerating the optimized function to be xml unparsed.
Use the reiterate() rule, rather than is_dead, to handle iteration.

Add a new test for DCE of side-effecting expressions. This meant adding the ability to add statements as well as removing them. We replace an assignment like "$x = f();", where $x is dead, with "f();". This requires a new ID for the new Eval_expr. next_id() will generate us a new node ID, but this is very much a hack. Counter starts from zero with each session, so we need to record where to start with our IDs, but that is future work.