Add test for set membership (Set::contains). Add Dead-code elimination (nearly). This doesn't remove remove statements yet, we'll need to change the interface for that. But it replaces $x = a(); with a();
The expands backwards-flow analysis to either forward- or backward-flow. It also allows a "full" analysis, where all statements are handled, or a "sparse" analysis, where only a small number of statements are looked at.

To complete the DCE, we need to allow CFG nodes to be replaced, to do simple alias analysis, and to convert the CFG back into a linearized MIR.