Support branches in the CFG. Add edge properties to indicate if the edge is a true branch, a false branch, or not a branch (using tribool for 3-state booleans).
Rather than numbering the vertices manually, change to renumbering before they are used. This fixes some bugs in the depth_first_search, which relies on the numbering. It has the constraint that each index is less than the number of vertices in the graph.

Also fixes a missing break with Index_array.