Add a test to InterpretOptimized, that is, an uppered dump of MIR in the generate-c pass. This is to faciliate the dataflow branch, where this will be used to test the optimizations.
After the lfc pass, __MAIN__ is created. When dumped, assignments to __MAIN__'s locals will not update global variables. Therefore, we need to remove statements from __MAIN__ and put them back into the global scope. This adds the Main_uppering pass to do that. (A better approach might be to replace variables with assignments to GLOBAL - comment added to that effect).