Allow Rvalues in the definition of Expr. This means that VARIABLE_NAME is in there twice, once as part of an RValue, once as part of a VARIABLE_NAME. I dont know if this will be a problem.
This allows Rvalues in phi args, which was the point.

When Phi nodes uses sets of arguments, the order didnt stay the same. Fixed.

Added more checks to the consistency check. It now checks edges are ok, and that they are a successor of their source, and a predecessor of their target.

CFG::replace_edge now operates directly on the (Rvalue, edge) pair. Using a foreach loop means using a copy of the pair, which then doesnt update the original. There are still bugs with this, but I cant find them.