The definition of SSA_edge was quite off. It combined two ideas: firstly, the variable used in the use-def chain, secondly, the location of the use/def. This is split now into two concepts, the latter being the SSA_op. An SSA_op is an operation on an VARIABLE, and can be an SSA_formal, SSA_stmt, SSA_branch or SSA_phi. SSA_edge is now a variable and the SSA_op.
For each defining/using statement in an SSA_op, there is not a single SSA_op. Instead, SSA_ops can be used to index maps using ssa_op_ptr_comparison. In DCE, these are used to mark statements as dead.

This also allows the def_use chains functions to be more precise. Sometimes, you want just the Phi nodes in a BB, sometimes just the statement. Def_use now provides a wider variety of specific functions that return either SSA_ops, or VARIABLE_NAMES (or lists). Rather than messing about at the call-site of these functions, its easier to define a function which returns the specific thing requested (I got into trouble before, trying to re-use pre_ssa_bb_uses for other purposes - it was difficult to tell if it was what I wanted, principally as a result of getting confused over what happened with phis).

This compiles, but doesnt link: a number of functions remain to be written.