Finish the majority of the code generator, including a few problems with the macros.
The body now instantiates properly, including interpolation and macro calls.
There is a slight bit of weak-typing going on, but its not allowed in interpolation, as that led to a few errors when attributes weren't defined.
phc_internal_node can now take an MICG node for line number and filename info. Ill have to add column names soon.

I dont actually want to use the value of a VARIABLE_NAME for variables in the generated code, so change the signature of write_var et al.
I've started using a pseudo-hungarian notation for Tokens, when the appropriate term is already used by a string.