Simple programs are compiling again :)
Add callbacks. They are currently limited to taking a single string as their parameter, but thats fine for the two callbacks we have. This switches the weird parameters (.hash and .length) to use callbacks, which is much nicer than defining 'default' properties.

Add column numbers into errors.
Add filename to MICG, which appears in the error message.
Make ; optional for callbacks and macro_calls.
Fix bug where only the first rule was evaluated if it was true.