Update for 0.2.0.1 release.
This adds news, and updates links and version numbers.

I also took the chance to reduce redundancy, by moving all files to src/archive, and not storing them anywhere else (I think this was always done with source until I broke it, but I've extended it to doc files too).

Instead of symlinks, use redirect rules everywhere. This puts all the links in one place, and means we dont have to rely on symlinks which arent in the repository.