While trying to get the type-inference to annotate all the nodes in the program, I got distracted by the problem of flow-sensitivity, which renders some of the work thus far a little bit off. So a few of the functions involved have been replaced by phc_TODOs. However, much of the restructuring for running the annotation part of the analyses is done.
This creates an Annotator, which is simply a CFG_visitor which runs before and after each statement is processed by the points-to analysis. You can (and probably should) have different annotators to run before and after the statements.

Analyses are split into "first class citizens" and "second class citizens". Its silly to pretend that the Points-to analysis, Type-analysis and CCP arent the most important analyses, or to try to generalize their jobs so that the other analyses can use the interface (which they wont).