Change the interface for WPA analyses. This one is clearer, and provides a way to differentiate between an assign_by_ref and an assign_by_copy. The intention is to provide all the interfaces that any optimizations need, and let them be unimplemented, rather than the current haphazard scheme.
An important part is that it adds kill_value and kill_reference. Before, it was unlcear what set_reference and set_value meant if an existing value was there. In some cases, it would be killing, and in some cases it would not. The CERTAINTY parameter to set_reference and set_value can be ignored in most cases, since kill_value and kill_reference should accomplish that. However, I'm not ready to get rid of it just yet.

The branch compiles, but breaks as soon as it touches the optimizer.

The next step is to fix all the optimizations to use this