It was incorrect to put all of the function's use/defs into the exit block. This reverts to the old method of keeping them in a set for the function. However, wht was correct was related: for defs of non-local variables, we should put uses of the in the exit block. While this is convervative (it may prevent a statement from being killed which we could safely kill), I would need to propagate liveness backwards to do that, and thats a little tricky.
This also tries to get DCE working again. The def-use-web is still teetering between the two representations. This moves the "old" use_def_chains and def_use_chains to the "new" representation. They are necessary. However, the differentiation between an SSA_STMT, and SSA_FORMAL and an SSA_BRANCH is not. This replaces all of those with a single SSA_BB. I think this means we wont need get_block_*.

The idea is to continue through DCE, re-enabling and fixing the Def_use_web. Once DCE is finished, I can kill the unused parts, and focus on making the used parts saner.