is_pure returned is_side_effecting, when it should have been the opposite. Handle literals in actual parameters. Record the uses of the operands of bin-ops. Inlining triggered in the wrong case. It should do it when there _is_ a user function.
The Whole_program analysis wouldnt converge, because the solutions were different. This was because I created a new basic_block for summary results in every iteration. This basic_block had a different index every time, so the results were never considered to be the same. I now re-use the first oe every time, and the results typically converge quickly, as expected.