Switch to context-sensitive analysis everywhere. It's not really possible to do this in small chunks, unfortunately. Basically, everywhere we used a Basic_block, we now use a context.
The context is made up of a list of basic_blocks, which is the call-string up to that point. It can be converted to a string, which is the BB ids, joined with '/'.

Contexts are used to provide symtable names, and you can build a context as a 'peer' - given a context from another basic block - or as a child, given the parent basic block.

Contexts can be either contextual, meaning the context matters, or non-contextual, meaning that all the contexts have been combined into a single result. The latter is not yet implemented, but will need to be implemented before SSA creation. There are hooks in place for this already.

This provides an 'infinite' call-string, which means it will break for recursion. I'll fix that soon.

This isnt tested, but it compiles and there is little that is functionally different, so only minor fallout is expected.

We stash a context to allow visit_block to work in Whole_program.cpp. However, the stash will be overwritten with function calls, so we need to fix this. The simplest way might be to simply have one context for all of Whole_porgram, and push and pop from it as we go.