Adds a result_state (either IN, OUT, WORKING or POSTBIND), representing the state in the basic block. Traditionally, compiler analyses use IN, OUT, GEN and KILL or something similar. However, there are many steps involved in evaluating an IR statement, so its not really straightforward to model them as equations to be solved. Instead, we apply the updates to WORKING as we go, and at the end of a statement copy them to OUT. POSTBIND stores the result of a forward/backward bind, which is copied into WORKING afterwards.
The main advantage is that by making the state explicit, we avoid propagating
result from the wrong result set.  For example, forward_bind copied from IN
before, but it should have copied from WORKING (which is the problem I was trying to solve).

The problem this leads to is that there are too many clones, and therefore the results take forever. I'll get to that later I suppose.

(This also fixes a minor problem in Bin_op - which I found trying to diagnose a different problem).