Fix bug reported by Alberto Bacchelli.
parent:: is used, even though there is no parent class. This will only trigger on buggy code, however, in the case of the reporter, this code wouldn't hit at run-time. As a result I went with a warning, and leaving the construct as is. A better solution would be to replace it with a run-time warning, but that's not really easy to do.

The error message lacked line number and file information, so I fixed a few places in the parser to add these. No new errors occur as a result.